Reorganize service module
