+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/service/mod.rs
@@ -1,10 +1,14 @@
 use cdk_framework::act::node::candid::Service;
+use swc_ecma_ast::ClassDecl;
 
 use crate::ts_ast::TsAst;
+use crate::ts_ast::{source_map::SourceMapped, GetName};
 use get_service_class_decls::GetFlattenedServiceClassDecls;
+use vm_value_conversions::{from_vm_value, list_from_vm_value, list_to_vm_value, to_vm_value};
 
 mod get_service_class_decls;
-mod to_service;
+mod member;
+mod vm_value_conversions;
 
 impl TsAst {
     pub fn build_services(&self) -> Vec<Service> {
@@ -14,5 +18,21 @@ impl TsAst {
             .iter()
             .map(|service_class_decl| service_class_decl.to_service())
             .collect()
+    }
+}
+
+impl SourceMapped<'_, ClassDecl> {
+    pub fn to_service(&self) -> Service {
+        let name = self.ident.get_name().to_string();
+        let methods = self.build_service_methods();
+
+        Service {
+            name,
+            methods,
+            to_vm_value,
+            list_to_vm_value,
+            from_vm_value,
+            list_from_vm_value,
+        }
     }
 }
