+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/mod.rs
@@ -18,7 +18,6 @@ mod ts_type_ann;
 pub mod azle_program;
 pub mod azle_type;
 pub mod azle_type_alias_decl;
-pub mod class_prop;
 pub mod param;
 pub mod source_map;
 pub mod traits;
